feat(gaussdb-a): add SONAR & EDC_RADAR oracles, fix schema/visitor/QPS issues
- Register SONAR oracle in GaussDBAOracleFactory
- Add EDC_RADAR oracle (GaussDBAEDCRadar + GaussDBAEDCRadarOracle)
- Fix GaussDBASchema.fromConnection() to exclude public schema tables
- Fix GaussDBAUpdateGenerator/DeleteGenerator to use getRandomTableOrBailout()
- Add GaussDBAPostfixText and GaussDBAManuelPredicate visitors to GaussDBAToStringVisitor
- Fix QPS display precision in Main.java (%d -> %.1f)
- Update release notes to v2.7.2
- GaussDB-A now supports 25 oracles (aligned with PostgreSQL)

Integration tested: SONAR 120 queries, EDC_RADAR 54 queries, TLP_WHERE 115 queries

+++ src/sqlancer/Main.java
@@ -911,8 +911,8 @@ public final class Main {
                 DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
                 Date date = new Date();
                 System.out.println(String.format(
-                        "[%s] Executed %d queries (%d queries/s; %.2f/s dbs, successful statements: %2d%%). Threads shut down: %d.",
-                        dateFormat.format(date), currentNrQueries, (int) throughput, throughputDbs,
+                        "[%s] Executed %d queries (%.1f queries/s; %.2f/s dbs, successful statements: %2d%%). Threads shut down: %d.",
+                        dateFormat.format(date), currentNrQueries, throughput, throughputDbs,
                         successfulStatementsRatio, threadsShutdown.get()));
                 timeMillis = System.currentTimeMillis();
                 lastNrQueries = currentNrQueries;
